# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prime Packaging Boxes API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithProducts"
        "404":
          description: Not found

  # ── Blog ──────────────────────────────────────────────────────────────────
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories (admin)
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    put:
      operationId: updateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin Products CRUD ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        categorySlug:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isShowcase:
          type: boolean
        isActive:
          type: boolean
        minOrder:
          type: integer
          nullable: true
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, slug, isFeatured, isActive, sortOrder, images]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: string
          nullable: true
        metaDescription:
          type: string
          nullable: true
        productCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, slug, isActive, isFeatured, sortOrder]
    CategoryWithProducts:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: "#/components/schemas/Product"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
      required: [success]
...
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
      required: [name, slug]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
...